🔵 Move Api documentation to another class
