+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/infrastructure/controllers/getproductcatalog/GetProductCatalogController.kt
@@ -4,55 +4,26 @@ import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.GetP
 import com.capitole.capitoleproductcatalogapi.domain.product.Category
 import com.capitole.capitoleproductcatalogapi.domain.product.SortField
 import com.capitole.capitoleproductcatalogapi.domain.product.SortOrder
-import io.swagger.v3.oas.annotations.Operation
-import io.swagger.v3.oas.annotations.Parameter
-import io.swagger.v3.oas.annotations.media.Schema
-import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-@Tag(name = "Products", description = "Retrieve, filter & sort product catalog")
-class GetProductCatalogController(private val getProductCatalog: GetProductCatalog) {
-  @Operation(
-      summary = "List products",
-      description = "Returns all products, with optional filtering by category and sorting"
-  )
+class GetProductCatalogController(
+  private val getProductCatalog: GetProductCatalog
+) : ProductCatalogApi {
+
   @GetMapping("/products", produces = [MediaType.APPLICATION_JSON_VALUE])
-  fun getProductCatalog(
-    @Parameter(
-        description = "Filter by category",
-        schema = Schema(
-            type = "string",
-            allowableValues = [
-              "ELECTRONICS",
-              "HOME_AND_KITCHEN",
-              "CLOTHING",
-              "ACCESSORIES",
-              "SPORTS",
-              "STATIONERY",
-              "TOYS_AND_GAMES",
-              "MUSICAL_INSTRUMENTS",
-              "FOOTWEAR",
-              "HOME_APPLIANCES"
-            ]
-        ), required = false
-    )
-    @RequestParam(name = "category", required = false) categoryParam: String?,
-    @Parameter(description = "Sort field: SKU, PRICE, DESCRIPTION or CATEGORY")
-    @RequestParam(required = false) sortField: SortField?,
-    @Parameter(description = "Sort order: ASC or DESC")
-    @RequestParam(required = false, defaultValue = "ASC") sortOrder: SortOrder
+  override fun getProductCatalog(
+    category: String?,
+    sortField: SortField?,
+    sortOrder: SortOrder
   ): ResponseEntity<GetProductCatalogResponse> {
-    val category = categoryParam?.let { Category.from(it) }
-    val getProductCatalogDTO = getProductCatalog.execute(category, sortField, sortOrder)
-    val getProductCatalogResponse = getProductCatalogDTO.toGetProductCatalogResponse()
+    val catEnum = category?.let { Category.from(it) }
+    val dto = getProductCatalog.execute(catEnum, sortField, sortOrder)
     return ResponseEntity.ok()
         .contentType(MediaType.APPLICATION_JSON)
-        .body(getProductCatalogResponse)
+        .body(dto.toGetProductCatalogResponse())
   }
 }
-
